Hoist createBumpStore + onBump into a shared frontend module
jobs.ts and maintenance.ts both needed the same cross-component "bump
to signal a refetch" primitive. jobs.ts had createBumpStore + onBump
defined inline; maintenance.ts open-coded the same pattern with a
writable + update(n => n+1).

Move both helpers to lib/stores/_bumpStore.ts. jobs.ts re-exports
onBump so existing imports of `$lib/stores/jobs` keep working.
maintenance.ts now destructures [maintenanceRefresh, bumpMaintenance
Refresh] from createBumpStore — one canonical home for the no-op
`n => n+1` quirk comment.
